requirements.txt
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
configobj==5.0.6
configparser==3.5.0
cryptography==2.3
decorator==4.3.0
distro==1.3.0
dj-database-url==0.5.0
Django==2.0.7
django-heroku==0.3.1
django-reset-migrations==0.3.1
django-tables2==1.21.2
dxfgrabber==1.0.0
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
isodate==0.6.0
lxml==4.2.3
mock==2.0.0
more-itertools==4.2.0
mysqlclient==1.3.13
networkx==2.1
nibabel==2.3.0
nipype==1.1.0
numpy==1.15.0
packaging==17.1
pandas==0.23.3
pbr==4.2.0
pdfminer.six==20170720
Pillow==5.2.0
pluggy==0.7.1
prov==1.5.0
py==1.5.4
pycparser==2.18
pycryptodome==3.6.4
pydot==1.2.4
pydotplus==2.0.2
PyMuPDF==1.13.15
pyOpenSSL==18.0.0
pyparsing==2.2.0
PyPDF2==1.26.0
pytest==3.6.4
python-dateutil==2.7.3
pytz==2018.5
rdflib==4.2.2
requests==2.19.1
scipy==1.1.0
simplejson==3.16.0
six==1.11.0
tabula-py==1.2.0
tabulate==0.8.2
traits==4.6.0
urllib3==1.23
whitenoise==3.3.1
psycopg2==2.7.2
